server/requirements.txt
flask
Werkzeug
python-dotenv
pymongo
flask-cors